refactor(Settings): split into modal and non-modal
