+++ src/components/SettingsModal/index.jsx
@@ -3,7 +3,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Modal from '@material-ui/core/Modal';
 import Backdrop from '@material-ui/core/Backdrop';
 import Fade from '@material-ui/core/Fade';
-import Settings from 'components/settings/Settings';
+import Settings from 'components/Settings';
 
 const SettingsModal = ({ settings, onSettings, onSave }) => {
   const useStyles = makeStyles(theme => ({
